refactor: move createLoopTask to startup and load empty file instead
